add:splash animation
