+++ app/src/main/java/com/stormers/storm/network/RetrofitClient.kt
@@ -8,7 +8,7 @@ import retrofit2.create
 object RetrofitClient {
     //Todo: URL 채우기
 
-    private const val BASE_URL = "http://52.78.113.197:3000"
+    private const val BASE_URL = "http://6cc8b8f248dd.ngrok.io"
 
     private fun getInstance() : Retrofit {
         return Retrofit.Builder()
