+++ app/src/main/res/raw/splash_real.json
@@ -1,0 +1,1 @@
+{"v":"5.7.0","fr":23.9759979248047,"ip":0,"op":47.9999958454548,"w":1125,"h":2436,"nm":"옴밈미-09","ddd":0,"assets":[{"id":"comp_0","layers":[{"ddd":0,"ind":1,"ty":4,"nm":"레이어 3 Outlines","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":0,"s":[0]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":13,"s":[100]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":31,"s":[100]},{"t":35.9999968840911,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":1,"k":[{"i":{"x":0.536,"y":1},"o":{"x":0.424,"y":0},"t":0,"s":[790.219,1136.767,0],"to":[0,0,0],"ti":[88.422,46.96,0]},{"i":{"x":0.648,"y":0.819},"o":{"x":0.597,"y":0},"t":14,"s":[589.904,1206.285,0],"to":[-62.136,-13.21,0],"ti":[11.725,4.603,0]},{"i":{"x":0.648,"y":0.815},"o":{"x":0.303,"y":0.16},"t":22,"s":[510.383,1170.271,0],"to":[-1.604,-0.63,0],"ti":[11.725,4.603,0]},{"t":22.9999980092804,"s":[510.383,1170.271,0]}],"ix":2},"a":{"a":0,"k":[61.19,69.189,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.833,0.833,0.833],"y":[0.833,0.833,0.833]},"o":{"x":[0.167,0.167,0.167],"y":[0.167,0.167,0.167]},"t":0,"s":[110,110,100]},{"t":17.9999984420456,"s":[100,100,100]}],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":1,"k":[{"i":{"x":0.861,"y":1},"o":{"x":0.564,"y":0},"t":11,"s":[{"i":[[0,-23.163],[23.162,0],[0,23.162],[-23.164,0]],"o":[[0,23.162],[-23.164,0],[0,-23.163],[23.162,0]],"v":[[41.94,0],[0.001,41.94],[-41.94,0],[0.001,-41.94]],"c":true}]},{"t":21.9999980958335,"s":[{"i":[[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0]],"v":[[188.628,15.9],[15.107,187.043],[-41.94,129.203],[131.58,-41.94]],"c":true}]}],"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.961000031116,0.760999971278,0.270999983245,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[42.19,42.19],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 1","np":2,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":47.9999958454548,"st":0,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"레이어 2 Outlines","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":0,"s":[0]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":13,"s":[100]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":31,"s":[100]},{"t":35.9999968840911,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":1,"k":[{"i":{"x":0.414,"y":1},"o":{"x":0.594,"y":0},"t":0,"s":[489.605,918.526,0],"to":[192.395,27.474,0],"ti":[0,0,0]},{"i":{"x":0.493,"y":1},"o":{"x":0.421,"y":0},"t":14,"s":[547.605,1112.526,0],"to":[0,0,0],"ti":[0,0,0]},{"t":22.9999980092804,"s":[549.605,1109.526,0]}],"ix":2},"a":{"a":0,"k":[69.588,69.587,0],"ix":1},"s":{"a":0,"k":[100,100,100],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":1,"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.613,"y":0},"t":8,"s":[{"i":[[0,-38.294],[38.294,0],[0,38.294],[-38.294,0]],"o":[[0,38.294],[-38.294,0],[0,-38.294],[38.294,0]],"v":[[69.338,0],[0,69.337],[-69.338,0],[0,-69.337]],"c":true}]},{"t":22.9999980092804,"s":[{"i":[[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0]],"v":[[119.231,-53.498],[-54.29,117.645],[-111.337,59.806],[62.182,-111.338]],"c":true}]}],"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.525490196078,0.427450980392,0.788235294118,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[69.587,69.588],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 1","np":2,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":47.9999958454548,"st":0,"bm":0},{"ddd":0,"ind":3,"ty":4,"nm":"레이어 4 Outlines","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":0,"s":[0]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":13,"s":[100]},{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":31,"s":[100]},{"t":35.9999968840911,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":1,"k":[{"i":{"x":0.546,"y":1},"o":{"x":0.556,"y":0},"t":0,"s":[380.268,1235.775,0],"to":[69.333,211.333,0],"ti":[0,0,0]},{"i":{"x":0.667,"y":1},"o":{"x":0.231,"y":0},"t":14,"s":[578.268,1323.775,0],"to":[0,0,0],"ti":[0,0,0]},{"t":22.9999980092804,"s":[567.268,1319.775,0]}],"ix":2},"a":{"a":0,"k":[111.318,111.317,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.667,0.667,0.667],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"t":14,"s":[81,81,100]},{"i":{"x":[0.667,0.667,0.667],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"t":22,"s":[100,100,100]},{"t":22.9999980092804,"s":[100,100,100]}],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":1,"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.587,"y":0},"t":8,"s":[{"i":[[0,-61.342],[61.341,0],[0,61.341],[-61.341,0]],"o":[[0,61.341],[-61.341,0],[0,-61.342],[61.341,0]],"v":[[111.068,0],[0,111.068],[-111.068,0],[0,-111.068]],"c":true}]},{"t":22.9999980092804,"s":[{"i":[[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0]],"v":[[119.5,-53.226],[-54.021,117.917],[-111.068,60.077],[62.452,-111.067]],"c":true}]}],"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.925,0.395999983245,0.395999983245,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[111.318,111.318],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 1","np":2,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":47.9999958454548,"st":0,"bm":0}]}],"layers":[{"ddd":0,"ind":1,"ty":0,"nm":"Pre-comp 1","refId":"comp_0","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":28,"s":[100]},{"t":29.9999974034093,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[562.5,1218,0],"ix":2},"a":{"a":0,"k":[562.5,1218,0],"ix":1},"s":{"a":0,"k":[100,100,100],"ix":6}},"ao":0,"w":1125,"h":2436,"ip":0,"op":31.9999972303032,"st":0,"bm":0},{"ddd":0,"ind":2,"ty":1,"nm":"White Solid 1","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":30,"s":[100]},{"t":31.9999972303032,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[562.5,1218,0],"ix":2},"a":{"a":0,"k":[562.5,1218,0],"ix":1},"s":{"a":0,"k":[100,100,100],"ix":6}},"ao":0,"sw":1125,"sh":2436,"sc":"#ffffff","ip":0,"op":31.9999972303032,"st":0,"bm":0}],"markers":[]}
